pyproject.toml
[project]
version = "1.0.0.0"

[tool.poetry]
name = "graphics"
version = "0.1.0"
description = ""
authors = ["Denix <densavochka@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.2.0"
build = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
